Move page containers next to pages
