+++ src/components/pages/berthPage/BerthPageContainer.tsx
@@ -14,7 +14,7 @@ import { LocalePush, withMatchParamsHandlers } from '../../../utils/container';
 import { BOAT_TYPES_BERTHS_QUERY } from '../../../utils/graphql';
 import { IconNames } from '../../common/Icon';
 import BoatsBerthsQuery from '../../query/BoatsBerthsQuery';
-import BerthPage from '../berthPage/BerthPage';
+import BerthPage from './BerthPage';
 
 import { Store } from '../../../redux/types';
 import { BerthFormValues } from '../../../types/berth';
